requirements.txt
slackclient>=2.0.0
flask
# requests